LayoutTests/imported/w3c/web-platform-tests/css/css-view-transitions/pseudo-get-computed-style-clean-style-expected.txt


FAIL properties of pseudo elements in update callback assert_equals: ::view-transition expected "absolute" but got "fixed"
